Implement copy only mode and some breaking changes

+++ src/helpers/clean.ts
@@ -1,7 +1,7 @@
 import fs from 'fs/promises'
 import path from 'path'
-import { log } from './log'
 import { getCopyFileConfigItems } from './tsbConfig'
+import { log } from './log'
 
 export const executeCleanFiles = async (
   tsbConfigPath: string
@@ -12,8 +12,14 @@ export const executeCleanFiles = async (
   const cwd = path.dirname(tsbConfigPath)
 
   for (const copyFilesConfigObject of copyfilesConfigObjects) {
-    const outDirectory = copyFilesConfigObject.outDirectory
-    log(`Removing directory ${outDirectory}`)
-    await fs.rm(path.join(cwd, outDirectory), { recursive: true, force: true })
+    if (copyFilesConfigObject.skipClean === true) {
+      log(`Skip removing directory for ${copyFilesConfigObject.outDirectories}`)
+      continue
+    }
+
+    for (const outDirectory of copyFilesConfigObject.outDirectories) {
+      log(`Removing directory ${outDirectory}`)
+      await fs.rm(path.join(cwd, outDirectory), { recursive: true, force: true })
+    }
   }
 }
